@@ -11,26 +11,28 @@
   },
   "dependencies": {
     "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/netinfo": "5.9.6",
-    "@react-navigation/native": "5.7.3",
-    "@react-navigation/stack": "5.9.0",
+    "@react-native-community/netinfo": "5.9.7",
+    "@react-navigation/native": "5.7.6",
+    "@react-navigation/stack": "5.9.3",
     "react": "16.13.1",
-    "react-native": "0.63.2",
-    "react-native-document-picker": "3.5.4",
+    "react-native": "0.63.3",
+    "react-native-document-picker": "4.0.0",
     "react-native-gesture-handler": "1.8.0",
-    "react-native-image-crop-picker": "0.34.1",
-    "react-native-reanimated": "1.13.0",
-    "react-native-safe-area-context": "3.1.7",
-    "react-native-screens": "2.10.1",
-    "stream-chat-react-native": "^"
+    "react-native-get-random-values": "1.5.0",
+    "react-native-image-crop-picker": "0.35.1",
+    "react-native-reanimated": "1.13.1",
+    "react-native-safe-area-context": "3.1.8",
+    "react-native-screens": "2.11.0",
+    "stream-chat": "2.6.0",
+    "stream-chat-react-native": "link:../../native-package"
   },
   "devDependencies": {
     "@babel/core": "7.11.6",
     "@babel/runtime": "7.11.2",
     "@react-native-community/eslint-config": "2.0.0",
-    "babel-jest": "26.3.0",
-    "eslint": "7.8.1",
-    "jest": "26.4.2",
+    "babel-jest": "26.5.2",
+    "eslint": "7.11.0",
+    "jest": "26.5.3",
     "metro-react-native-babel-preset": "0.63.0",
     "react-test-renderer": "16.13.1"
   },
