@@ -89,79 +88,90 @@
     }
   },
   "dependencies": {
-    "@babel/runtime": "^7.11.2",
-    "@stream-io/react-native-simple-markdown": "^1.2.1",
-    "@stream-io/styled-components": "^4.2.1",
+    "@babel/runtime": "7.11.2",
+    "@stream-io/react-native-simple-markdown": "1.2.1",
     "anchorme": "^1.1.2",
-    "babel-plugin-macros": "^2.8.0",
-    "dayjs": "^1.8.32",
-    "file-loader": "^3.0.1",
-    "i18next": "^19.3.1",
-    "lodash-es": "^4.17.11",
-    "metro-react-native-babel-preset": "^0.53.1",
-    "mime-types": "^2.1.24",
-    "path": "^0.12.7",
+    "dayjs": "1.9.3",
+    "file-loader": "6.1.1",
+    "i18next": "19.8.2",
+    "lodash-es": "4.17.15",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "mime-types": "2.1.27",
+    "path": "0.12.7",
     "react-art": "^16.8.6",
-    "react-native-actionsheet": "^2.4.2",
-    "react-native-image-zoom-viewer": "^3.0.1",
-    "seamless-immutable": "^7.1.4",
-    "stream-chat": "^2.2.2",
-    "uuid": "^3.4.0"
+    "react-native-actionsheet": "2.4.2",
+    "react-native-image-zoom-viewer": "3.0.1",
+    "seamless-immutable": "7.1.4",
+    "stream-chat": "2.6.0",
+    "styled-components": "5.2.0",
+    "uuid": "8.3.1"
   },
   "peerDependencies": {
     "react": "^15.0.0 || ^16.0.0",
     "react-dom": "^15.0.0 || ^16.0.0",
-    "react-native": "^0.57.0"
+    "react-native": "^0.60.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.4.3",
-    "@babel/polyfill": "^7.4.4",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/preset-flow": "^7.10.4",
-    "@babel/preset-react": "^7.0.0",
-    "@react-native-community/bob": "^0.16.2",
+    "@babel/cli": "7.11.6",
+    "@babel/core": "7.11.6",
+    "@babel/plugin-proposal-object-rest-spread": "7.11.0",
+    "@babel/polyfill": "7.11.5",
+    "@babel/preset-env": "7.11.5",
+    "@babel/preset-typescript": "7.10.4",
+    "@react-native-community/bob": "0.16.2",
     "@react-native-community/eslint-config": "2.0.0",
     "@react-native-community/eslint-plugin": "1.1.0",
-    "@testing-library/jest-native": "^3.3.0",
-    "@testing-library/react-native": "^7.0.1",
-    "@types/eslint": "^7.2.0",
-    "@types/react": "^16.8.23",
-    "@types/react-native": "^0.57.0",
-    "@typescript-eslint/eslint-plugin": "^3.7.1",
-    "@typescript-eslint/parser": "^3.7.1",
-    "ast-pretty-print": "^2.0.1",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-i18next-extract": "^0.6.1",
-    "babel-plugin-react-native-web": "^0.11.2",
-    "babel-preset-expo": "^5.0.0",
-    "eslint": "^7.6.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-better-styled-components": "^1.1.2",
-    "eslint-plugin-jest": "^23.20.0",
-    "eslint-plugin-markdown": "^1.0.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-sort-destructure-keys": "^1.3.5",
-    "eslint-plugin-typescript-sort-keys": "^1.3.0",
-    "husky": "^2.1.0",
-    "jest": "^26.2.2",
-    "mockdate": "^3.0.2",
-    "prettier": "^1.16.4",
-    "react": "^16.5.0",
-    "react-dom": "^16.8.6",
+    "@react-native-community/netinfo": "5.9.7",
+    "@testing-library/jest-native": "3.4.3",
+    "@testing-library/react-native": "7.0.2",
+    "@types/eslint": "7.2.4",
+    "@types/jest": "26.0.14",
+    "@types/lodash": "4.14.162",
+    "@types/mime-types": "2.1.0",
+    "@types/react": "16.9.52",
+    "@types/react-native": "0.63.25",
+    "@types/react-native-actionsheet": "2.4.1",
+    "@types/react-test-renderer": "16.9.3",
+    "@types/styled-components": "5.1.4",
+    "@typescript-eslint/eslint-plugin": "4.4.1",
+    "@typescript-eslint/parser": "4.4.1",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "26.5.2",
+    "babel-loader": "8.1.0",
+    "babel-plugin-i18next-extract": "0.8.0",
+    "babel-plugin-module-resolver": "4.0.0",
+    "babel-plugin-react-native-web": "0.14.1",
+    "eslint": "7.11.0",
+    "eslint-config-prettier": "6.12.0",
+    "eslint-plugin-babel": "5.3.1",
+    "eslint-plugin-better-styled-components": "1.1.2",
+    "eslint-plugin-jest": "24.1.0",
+    "eslint-plugin-markdown": "1.0.2",
+    "eslint-plugin-prettier": "3.1.4",
+    "eslint-plugin-react": "7.21.4",
+    "eslint-plugin-sort-destructure-keys": "1.3.5",
+    "eslint-plugin-typescript-sort-keys": "1.5.0",
+    "husky": "4.3.0",
+    "jest": "26.5.3",
+    "mockdate": "3.0.2",
+    "moment": "2.29.1",
+    "prettier": "2.1.2",
+    "react": "16.13.1",
+    "react-docgen-typescript": "1.20.5",
+    "react-dom": "16.13.1",
     "react-native": "0.61.5",
-    "react-native-web": "^0.11.4",
-    "react-styleguidist": "^8.0.6",
-    "react-test-renderer": "^16.13.1",
-    "stylelint": "^10.0.1",
-    "stylelint-config-standard": "^18.3.0",
-    "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.8.0",
-    "typescript": "^3.9.7",
-    "webpack": "4.19.1"
+    "react-native-typescript-transformer": "1.2.13",
+    "react-native-web": "0.14.1",
+    "react-styleguidist": "11.1.0",
+    "react-test-renderer": "16.13.1",
+    "stylelint": "13.7.2",
+    "stylelint-config-standard": "20.0.0",
+    "stylelint-config-styled-components": "0.1.1",
+    "stylelint-processor-styled-components": "1.10.0",
+    "typescript": "4.0.3",
+    "webpack": "4.44.2"
+  },
+  "resolutions": {
+    "styled-components": "5.2.0"
   }
 }
