@@ -1,34 +1,31 @@
-import React, { PureComponent } from 'react';
-import { View, SafeAreaView, TouchableOpacity, Text } from 'react-native';
-
+/* eslint-disable react/display-name */
+import 'react-native-get-random-values';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  LogBox,
+  SafeAreaView,
+  Text,
+  TouchableOpacity,
+  View,
+} from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator, useHeaderHeight } from '@react-navigation/stack';
+import { enableScreens } from 'react-native-screens';
 import { StreamChat } from 'stream-chat';
 import {
-  Chat,
   Channel,
-  MessageList,
-  MessageInput,
   ChannelList,
-  Thread,
-  CloseButton,
   ChannelPreviewMessenger,
-  Streami18n
+  Chat,
+  MessageInput,
+  MessageList,
+  Streami18n,
+  Thread,
 } from 'stream-chat-expo';
 
-import { createAppContainer, createStackNavigator } from 'react-navigation';
+LogBox.ignoreAllLogs(true);
+enableScreens();
 
-const chatClient = new StreamChat('q95x9hkbyd6p');
-const userToken =
-  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoicm9uIn0.eRVjxLvd4aqCEHY_JRa97g6k7WpHEhxL7Z4K4yTot1c';
-  const user = {
-    id: 'ron',
-  };
-  
-const filters = { type: 'messaging', example: 'example-apps', members: { '$in': ['ron'] } };
-const sort = { last_message_at: -1 };
-const options = {
-  state: true,
-  watch: true
-}
 // Read more about style customizations at - https://getstream.io/chat/react-native-chat/tutorial/#custom-styles
 const theme = {
   avatar: {
@@ -47,177 +44,188 @@ const theme = {
   },
 };
 
+const chatClient = new StreamChat('q95x9hkbyd6p');
+const userToken =
+  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoicm9uIn0.eRVjxLvd4aqCEHY_JRa97g6k7WpHEhxL7Z4K4yTot1c';
+const user = {
+  id: 'ron',
+};
+
+const filters = {
+  example: 'example-apps',
+  members: { $in: ['ron'] },
+  type: 'messaging',
+};
+const sort = { last_message_at: -1 };
+const options = {
+  state: true,
+  watch: true,
+};
+
 /**
  * Start playing with streami18n instance here:
  * Please refer to description of this PR for details: https://github.com/GetStream/stream-chat-react-native/pull/150
  */
 const streami18n = new Streami18n({
-  language: 'en'
+  language: 'en',
 });
 
-class ChannelListScreen extends PureComponent {
-  static navigationOptions = () => ({
-    headerTitle: <Text style={{ fontWeight: 'bold' }}>Channel List</Text>,
-  });
+const ChannelListScreen = React.memo(({ navigation }) => {
+  const { setChannel } = useContext(AppContext);
+  return (
+    <SafeAreaView>
+      <Chat client={chatClient} i18nInstance={streami18n} style={theme}>
+        <View style={{ height: '100%', padding: 10 }}>
+          <ChannelList
+            filters={filters}
+            onSelect={(channel) => {
+              setChannel(channel);
+              navigation.navigate('Channel');
+            }}
+            options={options}
+            Preview={ChannelPreviewMessenger}
+            sort={sort}
+          />
+        </View>
+      </Chat>
+    </SafeAreaView>
+  );
+});
 
-  render() {
-    return (
-      <SafeAreaView>
-        <Chat client={chatClient} style={theme} i18nInstance={streami18n}>
-          <View style={{ display: 'flex', height: '100%', padding: 10 }}>
-            <ChannelList
-              filters={filters}
-              sort={sort}
-              options={options}
-              Preview={ChannelPreviewMessenger}
-              onSelect={(channel) => {
-                this.props.navigation.navigate('Channel', {
-                  channel,
-                });
+const ChannelScreen = React.memo(({ navigation }) => {
+  const { channel, setThread } = useContext(AppContext);
+  const headerHeight = useHeaderHeight();
+
+  return (
+    <SafeAreaView>
+      <Chat client={chatClient} i18nInstance={streami18n} style={theme}>
+        <Channel channel={channel} keyboardVerticalOffset={headerHeight}>
+          <View style={{ flex: 1 }}>
+            <MessageList
+              onThreadSelect={(thread) => {
+                setThread(thread);
+                navigation.navigate('Thread', { channelId: channel.id });
               }}
             />
+            <MessageInput />
           </View>
-        </Chat>
-      </SafeAreaView>
-    );
-  }
-}
+        </Channel>
+      </Chat>
+    </SafeAreaView>
+  );
+});
 
-class ChannelScreen extends PureComponent {
-  static navigationOptions = ({ navigation }) => {
-    const channel = navigation.getParam('channel');
-    return {
-      headerTitle: (
-        <Text style={{ fontWeight: 'bold' }}>{channel.data.name}</Text>
-      ),
-    };
-  };
+const ThreadScreen = React.memo(({ route }) => {
+  const { thread } = useContext(AppContext);
+  const [channel] = useState(
+    chatClient.channel('messaging', route.params.channelId),
+  );
+  const headerHeight = useHeaderHeight();
 
-  render() {
-    const { navigation } = this.props;
-    const channel = navigation.getParam('channel');
-
-    return (
-      <SafeAreaView>
-        <Chat client={chatClient} style={theme} i18nInstance={streami18n}>
-          <Channel client={chatClient} channel={channel}>
-            <View style={{ display: 'flex', height: '100%' }}>
-              <MessageList
-                onThreadSelect={(thread) => {
-                  this.props.navigation.navigate('Thread', {
-                    thread,
-                    channel: channel.id,
-                  });
-                }}
-              />
-              <MessageInput />
-            </View>
-          </Channel>
-        </Chat>
-      </SafeAreaView>
-    );
-  }
-}
-
-class ThreadScreen extends PureComponent {
-  static navigationOptions = ({ navigation }) => ({
-    headerTitle: <Text style={{ fontWeight: 'bold' }}>Thread</Text>,
-    headerLeft: null,
-    headerRight: (
-      <TouchableOpacity
-        onPress={() => {
-          navigation.goBack();
-        }}
-        style={{
-          backgroundColor: '#ebebeb',
-          width: 30,
-          height: 30,
-          marginRight: 20,
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          borderRadius: 20,
-        }}
-      >
-        <CloseButton />
-      </TouchableOpacity>
-    ),
-  });
-
-  render() {
-    const { navigation } = this.props;
-    const thread = navigation.getParam('thread');
-    const channel = chatClient.channel(
-      'messaging',
-      navigation.getParam('channel'),
-    );
-
-    return (
-      <SafeAreaView>
-        <Chat client={chatClient} i18nInstance={streami18n}>
-          <Channel
-            client={chatClient}
-            channel={channel}
-            thread={thread}
-            dummyProp="DUMMY PROP"
+  return (
+    <SafeAreaView>
+      <Chat client={chatClient} i18nInstance={streami18n}>
+        <Channel
+          channel={channel}
+          keyboardVerticalOffset={headerHeight}
+          thread={thread}
+        >
+          <View
+            style={{
+              flex: 1,
+              justifyContent: 'flex-start',
+            }}
           >
-            <View
-              style={{
-                display: 'flex',
-                height: '100%',
-                justifyContent: 'flex-start',
-              }}
-            >
-              <Thread thread={thread} />
-            </View>
-          </Channel>
-        </Chat>
-      </SafeAreaView>
-    );
-  }
-}
+            <Thread thread={thread} />
+          </View>
+        </Channel>
+      </Chat>
+    </SafeAreaView>
+  );
+});
 
-const RootStack = createStackNavigator(
-  {
-    ChannelList: {
-      screen: ChannelListScreen,
-    },
-    Channel: {
-      screen: ChannelScreen,
-    },
-    Thread: {
-      screen: ThreadScreen,
-    },
-  },
-  {
-    initialRouteName: 'ChannelList',
-  },
-);
+const Stack = createStackNavigator();
 
-const AppContainer = createAppContainer(RootStack);
+const AppContext = React.createContext();
 
-export default class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      clientReady: false
-    }
-  }
+export default () => {
+  const [channel, setChannel] = useState();
+  const [clientReady, setClientReady] = useState(false);
+  const [thread, setThread] = useState();
 
-  async componentDidMount() {
+  useEffect(() => {
+    const setupClient = async () => {
+      await chatClient.setUser(user, userToken);
 
-    await chatClient.setUser(
-      user,
-      userToken,
-    );
+      setClientReady(true);
+    };
 
-    this.setState({
-      clientReady: true
-    })
-  }
-  render() {
-    if (this.state.clientReady)
-      return <AppContainer />;
-    else return null;
-  }
-}
+    setupClient();
+  }, []);
+
+  return (
+    <NavigationContainer>
+      <AppContext.Provider value={{ channel, setChannel, setThread, thread }}>
+        {clientReady && (
+          <Stack.Navigator
+            initialRouteName='ChannelList'
+            screenOptions={{
+              cardStyle: { backgroundColor: 'white' },
+              headerTitleStyle: { alignSelf: 'center', fontWeight: 'bold' },
+            }}
+          >
+            <Stack.Screen
+              component={ChannelScreen}
+              name='Channel'
+              options={() => ({
+                headerBackTitle: 'Back',
+                headerRight: () => <></>,
+                headerTitle: channel.data.name,
+              })}
+            />
+            <Stack.Screen
+              component={ChannelListScreen}
+              name='ChannelList'
+              options={{ headerTitle: 'Channel List' }}
+            />
+            <Stack.Screen
+              component={ThreadScreen}
+              name='Thread'
+              options={({ navigation }) => ({
+                headerLeft: () => <></>,
+                headerRight: () => (
+                  <TouchableOpacity
+                    onPress={() => {
+                      navigation.goBack();
+                    }}
+                    style={{
+                      alignItems: 'center',
+                      justifyContent: 'center',
+                      marginRight: 20,
+                    }}
+                  >
+                    <View
+                      style={{
+                        alignItems: 'center',
+                        backgroundColor: 'white',
+                        borderColor: 'rgba(0, 0, 0, 0.1)',
+                        borderRadius: 3,
+                        borderStyle: 'solid',
+                        borderWidth: 1,
+                        height: 30,
+                        justifyContent: 'center',
+                        width: 30,
+                      }}
+                    >
+                      <Text>X</Text>
+                    </View>
+                  </TouchableOpacity>
+                ),
+              })}
+            />
+          </Stack.Navigator>
+        )}
+      </AppContext.Provider>
+    </NavigationContainer>
+  );
+};
