@@ -1,31 +1,8 @@
-/**
- * In Expo 34, all the modules such as image picker, document picker, permissions etc etc have been moved to their own packages.
- * And they have been removed from Expo package. But these packages don't work with Expo 32 sdk (because of linking issue).
- */
 import { registerNativeHandlers } from 'stream-chat-react-native-core';
 import NetInfo from '@react-native-community/netinfo';
-import { Constants } from 'react-native-unimodules';
-import * as Expo from 'expo';
-
-let ImagePicker;
-let DocumentPicker;
-let Permissions;
-let manifest = {};
-
-manifest = Constants.manifest;
-if (!manifest) {
-  manifest = Expo.Constants.manifest;
-}
-
-if (manifest.sdkVersion.split('.')[0] >= 33) {
-  ImagePicker = require('expo-image-picker');
-  Permissions = require('expo-permissions');
-  DocumentPicker = require('expo-document-picker');
-} else {
-  ImagePicker = Expo.ImagePicker;
-  Permissions = Expo.Permissions;
-  DocumentPicker = Expo.DocumentPicker;
-}
+import * as DocumentPicker from 'expo-document-picker';
+import * as ImagePicker from 'expo-image-picker';
+import * as Permissions from 'expo-permissions';
 
 registerNativeHandlers({
   NetInfo: {
@@ -65,17 +42,55 @@ registerNativeHandlers({
       });
     },
   },
-  pickImage: async () => {
-    await Permissions.askAsync(Permissions.CAMERA_ROLL);
+  pickDocument: async ({ maxNumberOfFiles }) => {
+    try {
+      const { type, ...rest } = await DocumentPicker.getDocumentAsync();
 
-    return await ImagePicker.launchImageLibraryAsync({
-      allowsEditing: false,
-      aspect: [4, 3],
-      //TODO: Decide what to do about it
-      quality: 0.2,
-    });
+      if (type === 'cancel') {
+        return {
+          cancelled: true,
+        };
+      }
+      return {
+        cancelled: false,
+        docs: [rest],
+      };
+    } catch (err) {
+      return {
+        cancelled: true,
+      };
+    }
   },
-  pickDocument: async () => await DocumentPicker.getDocumentAsync(),
+  pickImage: async ({ compressImageQuality = 0.2, maxNumberOfFiles }) => {
+    try {
+      const { status } = await Permissions.askAsync(Permissions.CAMERA_ROLL);
+      if (status !== 'granted') {
+        return {
+          cancelled: true,
+        };
+      }
+
+      const { cancelled, ...rest } = await ImagePicker.launchImageLibraryAsync({
+        allowsEditing: false,
+        aspect: [4, 3],
+        quality: compressImageQuality,
+      });
+
+      if (cancelled) {
+        return {
+          cancelled,
+        };
+      }
+      return {
+        cancelled: false,
+        images: [{ uri: rest.uri }],
+      };
+    } catch (err) {
+      return {
+        cancelled: true,
+      };
+    }
+  },
 });
 
 export * from 'stream-chat-react-native-core';
