@@ -6,13 +6,13 @@ const i18nDirectoryRelativePath = '../src/i18n/';
 const directoryPath = path.join(__dirname, i18nDirectoryRelativePath);
 let countMissingTranslations = 0;
 
-fs.readdir(directoryPath, function(err, files) {
+fs.readdir(directoryPath, function (err, files) {
   if (err) {
     return console.log('Unable to scan directory: ' + err);
   }
 
-  files.forEach(function(file) {
-    if (file === 'index.js') return;
+  files.forEach(function (file) {
+    if (file === 'index.js' || file === 'index.ts') return;
     // Do whatever you want to do with the file
     const data = require(i18nDirectoryRelativePath + file);
     const keys = Object.keys(data);
@@ -20,7 +20,6 @@ fs.readdir(directoryPath, function(err, files) {
       if (!data[key] || data[key] === '') {
         countMissingTranslations = countMissingTranslations + 1;
         console.error(
-          '\033[91m',
           'Missing translation for key "' + key + '" in "' + file + '"',
         );
       }
