@@ -1,9 +1,11 @@
 import React from 'react';
 import { fireEvent, render, waitFor } from '@testing-library/react-native';
-import uuidv4 from 'uuid/v4';
+import { v4 as uuidv4 } from 'uuid';
 
-import Attachment from '../Attachment';
-import AttachmentActions from '../AttachmentActions';
+import { Attachment } from '../Attachment';
+import { AttachmentActions } from '../AttachmentActions';
+
+import { ThemeProvider } from '../../../contexts/themeContext/ThemeContext';
 
 import {
   generateAttachmentAction,
@@ -15,8 +17,16 @@ import {
   generateImgurAttachment,
 } from 'mock-builders/generator/attachment';
 
-const getAttachmentComponent = (props) => <Attachment {...props} />;
-const getActionComponent = (props) => <AttachmentActions {...props} />;
+const getAttachmentComponent = (props) => (
+  <ThemeProvider>
+    <Attachment {...props} />;
+  </ThemeProvider>
+);
+const getActionComponent = (props) => (
+  <ThemeProvider>
+    <AttachmentActions {...props} />;
+  </ThemeProvider>
+);
 
 describe('Attachment', () => {
   it('should render File component for "audio" type attachment', async () => {
