@@ -15,11 +15,12 @@ import { generateMessage } from 'mock-builders/generator/message';
 import { generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import Message from '../../Message';
-import Chat from '../../../Chat/Chat';
-import Channel from '../../../Channel/Channel';
-import MessageInput from '../../../MessageInput/MessageInput';
-import MessageList from '../../../MessageList/MessageList';
+import { Message } from '../../Message';
+
+import { Chat } from '../../../Chat/Chat';
+import { Channel } from '../../../Channel/Channel';
+import { MessageInput } from '../../../MessageInput/MessageInput';
+import { MessageList } from '../../../MessageList/MessageList';
 
 describe('Message', () => {
   let channel;
@@ -57,7 +58,6 @@ describe('Message', () => {
   });
 
   it('renders the Message and MessageSimple components', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const { getByTestId } = renderMessage({ message });
@@ -69,13 +69,13 @@ describe('Message', () => {
   });
 
   it('opens the action sheet on long press', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const { getByTestId, queryAllByTestId } = renderMessage({ message });
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
       expect(queryAllByTestId('cancel-button')).toHaveLength(0);
     });
@@ -89,7 +89,6 @@ describe('Message', () => {
   });
 
   it('opens the action sheet and gives the option to add a reaction', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const { getByTestId, getByText, queryAllByTestId } = renderMessage({
@@ -98,6 +97,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
@@ -110,7 +110,6 @@ describe('Message', () => {
   });
 
   it('closes the action sheet on press of the cancel button', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const { getByTestId, queryAllByTestId } = renderMessage({
@@ -119,6 +118,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
@@ -137,7 +137,6 @@ describe('Message', () => {
   });
 
   it('toggles edit message ability based on `isMyMessage` prop', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const {
@@ -152,16 +151,20 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
     fireEvent(getByTestId('message-content-wrapper'), 'longPress');
 
     await waitFor(() => {
-      expect(queryAllByText('Edit Message')).toHaveLength(0);
+      expect(getByText('Edit Message')).toBeTruthy();
     });
 
-    const message2 = generateMessage({ user: { ...user, id: 'id' } });
+    const message2 = generateMessage({
+      updated_at: 'newMessage',
+      user: { ...user, id: 'newUser' },
+    });
 
     rerender(
       <Chat client={chatClient}>
@@ -174,13 +177,12 @@ describe('Message', () => {
     fireEvent(getByTestId('message-content-wrapper'), 'longPress');
 
     await waitFor(() => {
-      expect(getByText('Edit Message')).toBeTruthy();
+      expect(queryAllByText('Edit Message')).toHaveLength(0);
     });
   });
 
   it('edits a message when selected from the action sheet', async () => {
-    const user = generateUser();
-    const message = generateMessage({ user: { ...user, id: 'id' } });
+    const message = generateMessage({ user });
     const clientOnSpy = jest.spyOn(chatClient, 'updateMessage');
 
     const { getByTestId, queryAllByTestId, queryAllByText } = renderMessage({
@@ -189,6 +191,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
@@ -212,7 +215,6 @@ describe('Message', () => {
   });
 
   it('toggles delete message ability based on `isMyMessage` prop', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
 
     const {
@@ -227,16 +229,20 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
     fireEvent(getByTestId('message-content-wrapper'), 'longPress');
 
     await waitFor(() => {
-      expect(queryAllByText('Delete Message')).toHaveLength(0);
+      expect(getByText('Delete Message')).toBeTruthy();
     });
 
-    const message2 = generateMessage({ user: { ...user, id: 'id' } });
+    const message2 = generateMessage({
+      updated_at: 'newMessage',
+      user: { ...user, id: 'newUser' },
+    });
 
     rerender(
       <Chat client={chatClient}>
@@ -249,13 +255,12 @@ describe('Message', () => {
     fireEvent(getByTestId('message-content-wrapper'), 'longPress');
 
     await waitFor(() => {
-      expect(getByText('Delete Message')).toBeTruthy();
+      expect(queryAllByText('Delete Message')).toHaveLength(0);
     });
   });
 
   it('deletes a message when selected from the action sheet', async () => {
-    const user = generateUser();
-    const message = generateMessage({ user: { ...user, id: 'id' } });
+    const message = generateMessage({ user });
     const clientOnSpy = jest.spyOn(chatClient, 'deleteMessage');
 
     const { getByTestId, queryAllByTestId, queryAllByText } = renderMessage({
@@ -264,6 +269,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
@@ -281,7 +287,6 @@ describe('Message', () => {
   });
 
   it('calls the `onThreadSelect` prop from the action sheet to navigate into a thread', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
     const onThreadSelect = jest.fn();
     const MessageComponent = () => (
@@ -302,6 +307,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-simple-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-container')).toHaveLength(0);
     });
 
@@ -319,7 +325,6 @@ describe('Message', () => {
   });
 
   it('calls the `onLongPress` prop function if it exists', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
     const onLongPress = jest.fn();
 
@@ -338,7 +343,6 @@ describe('Message', () => {
   });
 
   it('renders a custom action sheet when the `ActionSheet` prop exists', async () => {
-    const user = generateUser();
     const message = generateMessage({ user });
     // eslint-disable-next-line
     const ActionSheet = React.forwardRef((props, ref) => (
@@ -352,6 +356,7 @@ describe('Message', () => {
 
     await waitFor(() => {
       expect(getByTestId('message-wrapper')).toBeTruthy();
+      expect(getByTestId('message-content-wrapper')).toBeTruthy();
       expect(queryAllByTestId('action-sheet-prop')).toHaveLength(0);
     });
 
