@@ -5,9 +5,9 @@ import { generateMessage } from 'mock-builders/generator/message';
 import { generateStaticUser, generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import MessageStatus from '../MessageStatus';
+import { MessageStatus } from '../MessageStatus';
 
-import Chat from '../../../Chat/Chat';
+import { Chat } from '../../../Chat/Chat';
 
 import { Streami18n } from '../../../../utils/Streami18n';
 
