@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text } from 'react-native';
-import { ThemeProvider } from '@stream-io/styled-components';
 import { cleanup, render, waitFor } from '@testing-library/react-native';
+import { ThemeProvider } from 'styled-components/native';
 
 import {
   generateMessage,
@@ -9,9 +9,9 @@ import {
 } from 'mock-builders/generator/message';
 import { generateStaticUser } from 'mock-builders/generator/user';
 
-import MessageTextContainer from '../MessageTextContainer';
+import { MessageTextContainer } from '../MessageTextContainer';
 
-import { defaultTheme } from '../../../../styles/theme';
+import { defaultTheme } from '../../../../styles/themeConstants';
 
 afterEach(cleanup);
 
