@@ -7,9 +7,9 @@ import { generateChannel } from 'mock-builders/generator/channel';
 import { generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import ActionSheetAttachment from '../ActionSheetAttachment';
+import { ActionSheetAttachment } from '../ActionSheetAttachment';
 
-import Chat from '../../Chat/Chat';
+import { Chat } from '../../Chat/Chat';
 
 describe('ActionSheetAttachment', () => {
   const clientUser = generateUser();
