@@ -7,9 +7,10 @@ import {
 } from 'mock-builders/generator/message';
 import { generateStaticUser, generateUser } from 'mock-builders/generator/user';
 
-import DateSeparator from '../DateSeparator';
+import { DateSeparator } from '../DateSeparator';
 
-import { TranslationContext } from '../../../context';
+import { ThemeProvider } from '../../../contexts/themeContext/ThemeContext';
+import { TranslationProvider } from '../../../contexts/translationContext/TranslationContext';
 import { Streami18n } from '../../../utils/Streami18n';
 
 afterEach(cleanup);
@@ -21,9 +22,11 @@ describe('DateSeparator', () => {
     const user = generateUser();
     const message = generateMessage({ user });
     const { queryByTestId } = render(
-      <TranslationContext.Provider value={translators}>
-        <DateSeparator message={message} />
-      </TranslationContext.Provider>,
+      <ThemeProvider>
+        <TranslationProvider value={translators}>
+          <DateSeparator message={message} />
+        </TranslationProvider>
+      </ThemeProvider>,
     );
 
     await waitFor(() => {
@@ -37,12 +40,14 @@ describe('DateSeparator', () => {
     const user = generateUser();
     const message = generateMessage({ user });
     const { getByText } = render(
-      <TranslationContext.Provider value={translators}>
-        <DateSeparator
-          formatDate={(date) => date}
-          message={{ ...message, date: 'Hello World' }}
-        />
-      </TranslationContext.Provider>,
+      <ThemeProvider>
+        <TranslationProvider value={translators}>
+          <DateSeparator
+            formatDate={(date) => date}
+            message={{ ...message, date: 'Hello World' }}
+          />
+        </TranslationProvider>
+      </ThemeProvider>,
     );
 
     await waitFor(() => {
@@ -56,9 +61,11 @@ describe('DateSeparator', () => {
     const user = generateStaticUser(0);
     const message = generateStaticMessage('Hello World', { user });
     const { toJSON } = render(
-      <TranslationContext.Provider value={translators}>
-        <DateSeparator message={{ ...message, date: message.created_at }} />
-      </TranslationContext.Provider>,
+      <ThemeProvider>
+        <TranslationProvider value={translators}>
+          <DateSeparator message={{ ...message, date: message.created_at }} />
+        </TranslationProvider>
+      </ThemeProvider>,
     );
 
     await waitFor(() => {
