@@ -20,12 +20,15 @@ import {
 import { generateStaticUser, generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import MessageList from '../MessageList';
+import { MessageList } from '../MessageList';
 
-import Channel from '../../Channel/Channel';
-import Chat from '../../Chat/Chat';
+import { Channel } from '../../Channel/Channel';
+import { Chat } from '../../Chat/Chat';
 
-import { ChatContext } from '../../../context';
+import {
+  ChatContext,
+  ChatProvider,
+} from '../../../contexts/chatContext/ChatContext';
 
 describe('MessageList', () => {
   afterEach(cleanup);
@@ -126,6 +129,31 @@ describe('MessageList', () => {
     });
   });
 
+  it('should render the EmptyStateIndicator when no messages loaded', async () => {
+    const user1 = generateUser();
+    const mockedChannel = generateChannel({
+      members: [generateMember({ user: user1 })],
+      messages: [],
+    });
+
+    const chatClient = await getTestClientWithUser({ id: 'testID' });
+    useMockedApis(chatClient, [getOrCreateChannelApi(mockedChannel)]);
+    const channel = chatClient.channel('messaging', mockedChannel.id);
+    await channel.query();
+
+    const { getByTestId } = render(
+      <Chat client={chatClient}>
+        <Channel channel={channel}>
+          <MessageList />
+        </Channel>
+      </Chat>,
+    );
+
+    await waitFor(() => {
+      expect(getByTestId('empty-state')).toBeTruthy();
+    });
+  });
+
   it('should render the is offline error', async () => {
     const user1 = generateUser();
     const mockedChannel = generateChannel({
@@ -142,11 +170,11 @@ describe('MessageList', () => {
       <Chat client={chatClient}>
         <ChatContext.Consumer>
           {(context) => (
-            <ChatContext.Provider value={{ ...context, isOnline: false }}>
+            <ChatProvider value={{ ...context, isOnline: false }}>
               <Channel channel={channel}>
                 <MessageList />
               </Channel>
-            </ChatContext.Provider>
+            </ChatProvider>
           )}
         </ChatContext.Consumer>
       </Chat>,
