@@ -1,5 +1,5 @@
 import React from 'react';
-import { ThemeProvider } from '@stream-io/styled-components';
+import { ThemeProvider } from 'styled-components/native';
 import { cleanup, render, waitFor } from '@testing-library/react-native';
 
 import {
@@ -8,10 +8,10 @@ import {
 } from 'mock-builders/generator/message';
 import { generateStaticUser } from 'mock-builders/generator/user';
 
-import MessageSystem from '../MessageSystem';
+import { MessageSystem } from '../MessageSystem';
 
-import { TranslationContext } from '../../../context';
-import { defaultTheme } from '../../../styles/theme';
+import { TranslationProvider } from '../../../contexts/translationContext/TranslationContext';
+import { defaultTheme } from '../../../styles/themeConstants';
 import { Streami18n } from '../../../utils/Streami18n';
 
 afterEach(cleanup);
@@ -23,9 +23,9 @@ describe('MessageSystem', () => {
     const message = generateMessage();
     const { queryByTestId } = render(
       <ThemeProvider theme={defaultTheme}>
-        <TranslationContext.Provider value={translators}>
+        <TranslationProvider value={translators}>
           <MessageSystem message={message} />
-        </TranslationContext.Provider>
+        </TranslationProvider>
       </ThemeProvider>,
     );
 
@@ -41,9 +41,9 @@ describe('MessageSystem', () => {
     const message = generateStaticMessage('Hello World', { user });
     const { toJSON } = render(
       <ThemeProvider theme={defaultTheme}>
-        <TranslationContext.Provider value={translators}>
+        <TranslationProvider value={translators}>
           <MessageSystem message={message} />
-        </TranslationContext.Provider>
+        </TranslationProvider>
       </ThemeProvider>,
     );
 
