@@ -5,11 +5,12 @@ import { cleanup, render, waitFor } from '@testing-library/react-native';
 import { generateStaticUser, generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import TypingIndicator from '../TypingIndicator';
+import { TypingIndicator } from '../TypingIndicator';
 
-import Chat from '../../Chat/Chat';
+import { Chat } from '../../Chat/Chat';
 
-import { ChannelContext, TranslationContext } from '../../../context';
+import { ChannelProvider } from '../../../contexts/channelContext/ChannelContext';
+import { TranslationProvider } from '../../../contexts/translationContext/TranslationContext';
 import { Streami18n } from '../../../utils/Streami18n';
 
 afterEach(cleanup);
@@ -29,11 +30,11 @@ describe('TypingIndicator', () => {
 
     const { getByTestId } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
     expect(t).toHaveBeenCalledWith(
@@ -59,11 +60,11 @@ describe('TypingIndicator', () => {
 
     const { getByTestId } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
     expect(t).toHaveBeenCalledWith('{{ user }} is typing...', {
@@ -85,11 +86,11 @@ describe('TypingIndicator', () => {
 
     const { getByTestId, queryByTestId } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator Avatar={View} />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
     expect(t).toHaveBeenCalledWith('{{ user }} is typing...', {
@@ -115,11 +116,11 @@ describe('TypingIndicator', () => {
 
     const { getByTestId } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator Avatar={View} />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
 
@@ -140,11 +141,11 @@ describe('TypingIndicator', () => {
 
     const { getByTestId, queryByTestId } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator Avatar={View} />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
 
@@ -168,11 +169,11 @@ describe('TypingIndicator', () => {
 
     const { toJSON } = render(
       <Chat client={chatClient}>
-        <TranslationContext.Provider value={{ t }}>
-          <ChannelContext.Provider value={{ typing }}>
+        <TranslationProvider value={{ t }}>
+          <ChannelProvider value={{ typing }}>
             <TypingIndicator />
-          </ChannelContext.Provider>
-        </TranslationContext.Provider>
+          </ChannelProvider>
+        </TranslationProvider>
       </Chat>,
     );
     await waitFor(() => {
