@@ -4,9 +4,9 @@ import { render, waitFor } from '@testing-library/react-native';
 import { generateUser } from 'mock-builders/generator/user';
 import { getTestClientWithUser } from 'mock-builders/mock';
 
-import ReactionList from '../ReactionList';
+import { ReactionList } from '../ReactionList';
 
-import Chat from '../../Chat/Chat';
+import { Chat } from '../../Chat/Chat';
 
 const defaultProps = {
   getTotalReactionCount: () => 5,
@@ -168,7 +168,7 @@ describe('ReactionList', () => {
     );
 
     await waitFor(() => {
-      expect(queryByTestId('reaction-list')).toBeTruthy();
+      expect(queryByTestId('reaction-list')).toBeFalsy();
       emojiTypes.forEach((emoji) => expect(queryByTestId(emoji)).toBeFalsy());
     });
 
