@@ -1,21 +1,208 @@
+/* eslint-disable sort-keys */
 /* eslint-env node */
 const path = require('path');
 const webpack = require('webpack');
 
 const notBabeledDeps = [
-  '@stream-io/react-native-simple-markdown',
-  'react-native-image-pan-zoom',
   'react-native-image-zoom-viewer',
+  'react-native-image-pan-zoom',
+  '@stream-io/react-native-simple-markdown',
+];
+
+const sections = [
+  {
+    components: ['src/components/Chat/Chat.tsx'],
+    exampleMode: 'collapse',
+    name: 'Top Level Component',
+    usageMode: 'expand',
+  },
+  {
+    components: ['src/components/Channel/Channel.tsx'],
+    exampleMode: 'collapse',
+    name: 'Channel Component',
+    usageMode: 'expand',
+  },
+  {
+    components: ['src/components/Thread/Thread.tsx'],
+    exampleMode: 'collapse',
+    name: 'Thread Component',
+    usageMode: 'expand',
+  },
+  {
+    components: [
+      'src/components/ChannelList/ChannelList.tsx',
+      'src/components/ChannelList/ChannelListMessenger.tsx',
+      'src/components/ChannelPreview/ChannelPreview.tsx',
+      'src/components/ChannelPreview/ChannelPreviewMessenger.tsx',
+    ],
+    exampleMode: 'collapse',
+    name: 'ChannelList Components',
+    usageMode: 'expand',
+  },
+  {
+    components: [
+      'src/components/MessageList/DateSeparator.tsx',
+      'src/components/MessageList/MessageList.tsx',
+      'src/components/MessageList/MessageNotification.tsx',
+      'src/components/MessageList/MessageSystem.tsx',
+      'src/components/MessageList/TypingIndicator.tsx',
+      'src/components/MessageList/TypingIndicatorContainer.tsx',
+    ],
+    exampleMode: 'collapse',
+    name: 'MessageList Components',
+    usageMode: 'expand',
+  },
+  {
+    components: [
+      'src/components/Message/Message.tsx',
+      'src/components/Message/MessageSimple/MessageSimple.tsx',
+      'src/components/Message/MessageSimple/MessageActionSheet.tsx',
+      'src/components/Message/MessageSimple/MessageAvatar.tsx',
+      'src/components/Message/MessageSimple/MessageContent.tsx',
+      'src/components/Message/MessageSimple/MessageReplies.tsx',
+      'src/components/Message/MessageSimple/MessageStatus.tsx',
+      'src/components/Message/MessageSimple/MessageTextContainer.tsx',
+      'src/components/Attachment/Attachment.tsx',
+      'src/components/Attachment/AttachmentActions.tsx',
+      'src/components/Attachment/Card.tsx',
+      'src/components/Attachment/FileAttachment.tsx',
+      'src/components/Attachment/FileAttachmentGroup.tsx',
+      'src/components/Attachment/Gallery.tsx',
+    ],
+    exampleMode: 'collapse',
+    name: 'Message Components',
+    usageMode: 'expand',
+  },
+  {
+    components: [
+      'src/components/AutoCompleteInput/AutoCompleteInput.tsx',
+      'src/components/MessageInput/ActionSheetAttachment.tsx',
+      'src/components/MessageInput/AttachButton.tsx',
+      'src/components/AutoCompleteInput/CommandsItem.tsx',
+      'src/components/MessageInput/FileUploadPreview.tsx',
+      'src/components/MessageInput/ImageUploadPreview.tsx',
+      'src/components/AutoCompleteInput/MentionsItem.tsx',
+      'src/components/MessageInput/MessageInput.tsx',
+      'src/components/MessageInput/SendButton.tsx',
+      'src/components/MessageInput/UploadProgressIndicator.tsx',
+    ],
+    exampleMode: 'collapse',
+    name: 'MessageInput Components',
+    usageMode: 'expand',
+  },
+  {
+    components: [
+      'src/components/Avatar/Avatar.tsx',
+      'src/components/CloseButton/CloseButton.tsx',
+      'src/components/IconBadge.tsx',
+      'src/components/IconSquare.tsx',
+      'src/components/KeyboardCompatibleView/KeyboardCompatibleView.tsx',
+      'src/components/Indicators/LoadingIndicator.tsx',
+      'src/components/Reaction/ReactionList.tsx',
+      'src/components/Reaction/ReactionPicker.tsx',
+      'src/components/Reaction/ReactionPickerWrapper.tsx',
+      'src/components/Spinner/Spinner.tsx',
+    ],
+    exampleMode: 'collapse',
+    name: 'Miscellaneous',
+    sections: [
+      {
+        content: 'src/components/Message/MessageSimple/utils/renderText.md',
+        name: 'renderText',
+      },
+      {
+        content: 'src/utils/Streami18n.md',
+        name: 'Streami18n',
+      },
+    ],
+    usageMode: 'expand',
+  },
+  {
+    exampleMode: 'collapse',
+    name: 'Contexts',
+    sections: [
+      {
+        content: 'src/contexts/chatContext/ChatContext.md',
+        name: 'ChatContext',
+      },
+      {
+        content: 'src/contexts/chatContext/withChatContext.md',
+        name: 'withChatContext',
+      },
+      {
+        content: 'src/contexts/channelContext/ChannelContext.md',
+        name: 'ChannelContext',
+      },
+      {
+        content: 'src/contexts/channelContext/withChannelContext.md',
+        name: 'withChannelContext',
+      },
+      {
+        content: 'src/contexts/keyboardContext/KeyboardContext.md',
+        name: 'KeyboardContext',
+      },
+      {
+        content: 'src/contexts/keyboardContext/withKeyboardContext.md',
+        name: 'withKeyboardContext',
+      },
+      {
+        content: 'src/contexts/messagesContext/MessagesContext.md',
+        name: 'MessagesContext',
+      },
+      {
+        content: 'src/contexts/messagesContext/withMessagesContext.md',
+        name: 'withMessagesContext',
+      },
+      {
+        content: 'src/contexts/suggestionsContext/SuggestionsContext.md',
+        name: 'SuggestionsContext',
+      },
+      {
+        content: 'src/contexts/suggestionsContext/withSuggestionsContext.md',
+        name: 'withSuggestionsContext',
+      },
+      {
+        content: 'src/contexts/threadContext/ThreadContext.md',
+        name: 'ThreadContext',
+      },
+      {
+        content: 'src/contexts/threadContext/withThreadContext.md',
+        name: 'withThreadContext',
+      },
+    ],
+    usageMode: 'expand',
+  },
+  {
+    exampleMode: 'collapse',
+    name: 'Custom UI Component Props',
+    sections: [
+      {
+        content: 'src/components/docs/AttachmentProps.md',
+        name: 'Attachment (in MessageSimple)',
+      },
+      {
+        content: 'src/components/docs/ListProps.md',
+        name: 'List (in ChannelList)',
+      },
+      {
+        content: 'src/components/docs/MessageProps.md',
+        name: 'Message (in MessageList)',
+      },
+      {
+        content: 'src/components/Message/MessageSimple/MessageText.md',
+        name: 'MessageText (in MessageSimple)',
+      },
+      {
+        content: 'src/components/docs/PreviewProps.md',
+        name: 'Preview (in ChannelList)',
+      },
+    ],
+    usageMode: 'expand',
+  },
 ];
 
 module.exports = {
-  assetsDir: 'src/assets',
-  compilerConfig: {
-    objectAssign: 'Object.assign',
-    transforms: {
-      dangerousTaggedTemplateString: true,
-    },
-  },
+  title: 'Stream Chat React Native - Docs',
   require: [
     '@babel/polyfill',
     path.join(
@@ -23,232 +210,86 @@ module.exports = {
       'src/styleguideComponents/register-react-native-web.js',
     ),
   ],
+  styleguideDir: 'docs',
+  assetsDir: 'src/assets',
+  sortProps: (props) => props,
+  serverPort: 6068,
+  compilerConfig: {
+    transforms: {
+      dangerousTaggedTemplateString: true,
+      moduleImport: false,
+    },
+    objectAssign: 'Object.assign',
+  },
   resolver(ast, recast) {
     return require('react-docgen').resolver.findAllExportedComponentDefinitions(
       ast,
       recast,
     );
   },
-  sections: [
+  propsParser: require('react-docgen-typescript').withCustomConfig(
+    './tsconfig.json',
     {
-      components: ['src/components/Chat/Chat.js'],
-      exampleMode: 'collapse',
-      name: 'Top Level Component',
-      usageMode: 'expand',
+      propFilter: { skipPropsWithoutDoc: true },
     },
-    {
-      components: ['src/components/Channel/Channel.js'],
-      exampleMode: 'collapse',
-      name: 'Channel Component',
-      usageMode: 'expand',
-    },
-    {
-      components: ['src/components/Thread/Thread.js'],
-      exampleMode: 'collapse',
-      name: 'Thread Component',
-      usageMode: 'expand',
-    },
-    {
-      components: [
-        'src/components/ChannelList/ChannelList.js',
-        'src/components/ChannelList/ChannelListMessenger.js',
-        'src/components/ChannelPreview/ChannelPreview.js',
-        'src/components/ChannelPreview/ChannelPreviewMessenger.js',
-      ],
-      exampleMode: 'collapse',
-      name: 'ChannelList Components',
-      usageMode: 'expand',
-    },
-    {
-      components: [
-        'src/components/MessageList/DateSeparator.js',
-        'src/components/MessageList/EventIndicator.js',
-        'src/components/MessageList/MessageList.js',
-        'src/components/MessageList/MessageNotification.js',
-        'src/components/MessageList/MessageSystem.js',
-        'src/components/MessageList/TypingIndicator.js',
-      ],
-      exampleMode: 'collapse',
-      name: 'MessageList Components',
-      usageMode: 'expand',
-    },
-    {
-      components: [
-        'src/components/Attachment/Attachment.js',
-        'src/components/Attachment/Card.js',
-        'src/components/Attachment/FileAttachmentGroup.js',
-        'src/components/Attachment/Gallery.js',
-        'src/components/Message/Message.js',
-        'src/components/Message/MessageSimple/index.js',
-        'src/components/Message/MessageSimple/MessageActionSheet.js',
-        'src/components/Message/MessageSimple/MessageAvatar.js',
-        'src/components/Message/MessageSimple/MessageContent.js',
-        'src/components/Message/MessageSimple/MessageReplies.js',
-        'src/components/Message/MessageSimple/MessageStatus.js',
-        'src/components/Message/MessageSimple/MessageTextContainer.js',
-      ],
-      exampleMode: 'collapse',
-      name: 'Message Components',
-      usageMode: 'expand',
-    },
-    {
-      components: [
-        'src/components/AutoCompleteInput/AutoCompleteInput.js',
-        'src/components/MessageInput/ActionSheetAttachment.js',
-        'src/components/MessageInput/AttachButton.js',
-        'src/components/AutoCompleteInput/CommandsItem.js',
-        'src/components/MessageInput/FileUploadPreview.js',
-        'src/components/MessageInput/ImageUploadPreview.js',
-        'src/components/AutoCompleteInput/MentionsItem.js',
-        'src/components/MessageInput/MessageInput.js',
-        'src/components/MessageInput/SendButton.js',
-        'src/components/MessageInput/UploadProgressIndicator.js',
-      ],
-      exampleMode: 'collapse',
-      name: 'MessageInput Components',
-      usageMode: 'expand',
-    },
-    {
-      components: [
-        'src/components/Avatar/Avatar.js',
-        'src/components/CloseButton/CloseButton.js',
-        'src/components/Attachment/FileIcon.js',
-        'src/components/IconBadge.js',
-        'src/components/IconSquare.js',
-        'src/components/KeyboardCompatibleView/KeyboardCompatibleView.js',
-        'src/components/Indicators/LoadingIndicator.js',
-        'src/components/Reaction/ReactionList.js',
-        'src/components/Reaction/ReactionPicker.js',
-        'src/components/Reaction/ReactionPickerWrapper.js',
-        'src/components/Spinner/Spinner.js',
-        'src/components/SuggestionsProvider/SuggestionsProvider.js',
-      ],
-      exampleMode: 'collapse',
-      name: 'Miscellaneous',
-      sections: [
-        {
-          content: 'src/components/docs/renderText.md',
-          name: 'renderText',
-        },
-        {
-          content: 'src/components/docs/Streami18n.md',
-          name: 'Streami18n',
-        },
-      ],
-      usageMode: 'expand',
-    },
-    {
-      exampleMode: 'collapse',
-      name: 'Contexts',
-      sections: [
-        {
-          content: 'src/components/docs/ChatContext.md',
-          name: 'ChatContext',
-        },
-        {
-          content: 'src/components/docs/withChatContext.md',
-          name: 'withChatContext',
-        },
-        {
-          content: 'src/components/docs/ChannelContext.md',
-          name: 'ChannelContext',
-        },
-        {
-          content: 'src/components/docs/withChannelContext.md',
-          name: 'withChannelContext',
-        },
-        {
-          content: 'src/components/docs/KeyboardContext.md',
-          name: 'KeyboardContext',
-        },
-        {
-          content: 'src/components/docs/withKeyboardContext.md',
-          name: 'withKeyboardContext',
-        },
-        {
-          content: 'src/components/docs/SuggestionsContext.md',
-          name: 'SuggestionsContext',
-        },
-        {
-          content: 'src/components/docs/withSuggestionsContext.md',
-          name: 'withSuggestionsContext',
-        },
-      ],
-      usageMode: 'expand',
-    },
-    {
-      exampleMode: 'collapse',
-      name: 'Custom UI Component Props',
-      sections: [
-        {
-          content: 'src/components/docs/AttachmentProps.md',
-          name: 'Attachment (in MessageSimple)',
-        },
-        {
-          content: 'src/components/docs/ListProps.md',
-          name: 'List (in ChannelList)',
-        },
-        {
-          content: 'src/components/docs/MessageProps.md',
-          name: 'Message (in MessageList)',
-        },
-        {
-          content: 'src/components/docs/MessageText.md',
-          name: 'MessageText (in MessageSimple)',
-        },
-        {
-          content: 'src/components/docs/PreviewProps.md',
-          name: 'Preview (in ChannelList)',
-        },
-      ],
-      usageMode: 'expand',
-    },
-  ],
-  serverPort: 6068,
-  sortProps: (props) => props,
+  ).parse,
   styleguideComponents: {
     PathlineRenderer: path.join(
       __dirname,
-      'src/styleguideComponents/PathlineRenderer.js',
+      'src/styleguideComponents/PathlineRenderer.tsx',
     ),
-    Slot: path.join(__dirname, 'src/styleguideComponents/Slot.js'),
+    Wrapper: path.join(__dirname, 'src/styleguideComponents/Wrapper.tsx'),
   },
-  styleguideDir: 'docs',
+  sections,
   template: {
     favicon: 'https://getstream.imgix.net/images/favicons/favicon-96x96.png',
   },
-  title: 'Stream Chat React Native - Docs',
   webpackConfig: {
+    devtool: 'source-map',
+    resolve: {
+      // auto resolves any react-native import as react-native-web
+      alias: {
+        'react-native': 'react-native-web',
+        'react-native-gesture-handler': 'react-native-web',
+        'styled-components/native':
+          'styled-components/native/dist/styled-components.native.cjs.js',
+        'styled-components':
+          'styled-components/native/dist/styled-components.native.cjs.js',
+        // Looks ugly in docs, better to just not show it for now
+        'react-native-actionsheet': path.join(
+          __dirname,
+          'src/styleguideComponents/ReactNativeActionSheet.tsx',
+        ),
+      },
+      extensions: ['.web.js', '.js', '.ts', '.tsx'],
+    },
     devServer: {
       clientLogLevel: 'warn',
-    },
-    devtool: 'source-map',
-    externals: {
-      react: 'React',
-      'react-native': 'react-native',
-      'react-native-web-player': 'react-native-web-player',
     },
     module: {
       rules: [
         {
+          test: /\.(js|jsx|ts|tsx)$/,
+          loader: 'babel-loader',
           include: [
             path.join(__dirname, 'src'),
             ...notBabeledDeps.map((dep) =>
               path.join(__dirname, 'node_modules', dep),
             ),
           ],
-          loader: 'babel-loader',
           options: {
-            babelrc: false,
+            comments: true,
             plugins: [
-              'macros',
-              '@babel/plugin-transform-runtime',
-              '@babel/proposal-object-rest-spread',
+              'module-resolver',
               'react-native-web',
+              'babel-plugin-styled-components',
             ],
-            presets: ['@babel/env', 'module:metro-react-native-babel-preset'],
+            presets: [
+              'module:metro-react-native-babel-preset',
+              '@babel/preset-typescript',
+            ],
+            babelrc: false,
           },
-          test: /\.js$/,
         },
         {
           test: /\.(jpe?g|png|gif|ttf)$/i,
@@ -256,8 +297,8 @@ module.exports = {
             {
               loader: 'file-loader',
               options: {
-                digest: 'hex',
                 hash: 'sha512',
+                digest: 'hex',
                 name: '[hash].[ext]',
                 outputPath: 'build/images',
               },
@@ -273,20 +314,10 @@ module.exports = {
         __DEV__: process.env,
       }),
     ],
-    resolve: {
-      // auto resolves any react-native import as react-native-web
-      alias: {
-        '@stream-io/styled-components':
-          '@stream-io/styled-components/native/dist/styled-components.native.cjs.js',
-        'react-native': 'react-native-web',
-        // Looks ugly in docs, better to just not show it for now
-        'react-native-actionsheet': path.join(
-          __dirname,
-          'src/styleguideComponents/ReactNativeActionSheet.js',
-        ),
-        'react-native-gesture-handler': 'react-native-web',
-      },
-      extensions: ['.web.js', '.js'],
+    externals: {
+      react: 'React',
+      'react-native': 'react-native',
+      'react-native-web-player': 'react-native-web-player',
     },
   },
 };
